requirements.txt
mmcv==0.2.13
Polygon3==3.0.8
editdistance